Move bar plot aggregation back-end test to test_search.py
